address lloyds comments
