+++ src/pyvale/vfm/__init__.py
@@ -3,3 +3,33 @@
 # License: MIT
 # Copyright (C) 2025 The Computer Aided Validation Team
 #===============================================================================
+
+from .vfm import *
+
+from .experimentdata import *
+from .identification import *
+
+from .constlaw import *
+from .constlaws import *
+from .constparam import *
+from .hardening import *
+from .radialreturn import *
+
+from .spatialparam import *
+from .spatialparamknown import *
+from .spatialparamhomogeneous import *
+from .spatialparambasisfuncs import *
+
+from .optimiser import *
+from .optimiserleastsquares import *
+
+from .metric import *
+from .metricsbvf import *
+from .vfmesh import *
+
+from .objectivefunc import *
+from .objectivefuncscalar import *
+from .objectivefuncvector import *
+
+from .dof import *
+from .normalisation import *
