+++ src/pyvale/vfm/metric.py
@@ -3,11 +3,9 @@ from abc import ABC, abstractmethod
 import numpy as np
 import numpy.typing as npt
 
-from pyvale.vfm.constitutive_laws.constitutive_law import IConstitutiveLaw
-from pyvale.vfm.experiment_data import ExperimentData
-from pyvale.vfm.spatial_parameterisations.spatial_parameterisation import (
-    ISpatialParameterisation,
-)
+from pyvale.vfm.constlaw import IConstitutiveLaw
+from pyvale.vfm.experimentdata import ExperimentData
+from pyvale.vfm.spatialparam import ISpatialParameterisation
 
 
 class IMetric(ABC):
