+++ src/pyvale/vfm/objectivefuncvector.py
@@ -1,8 +1,8 @@
 import numpy as np
 import numpy.typing as npt
-from pyvale.vfm.objective_functions.objective_function import (
-    IVectorObjectiveFunction,
-)
+
+from pyvale.vfm.objectivefunc import IVectorObjectiveFunction
+
 
 class VectorFirstResultPassthrough(IVectorObjectiveFunction):
     def evaluate(
