+++ src/pyvale/vfm/spatialparambasisfuncs.py
@@ -1,20 +1,19 @@
 from abc import ABC, abstractmethod
 from copy import copy
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 
 import numpy as np
 import numpy.typing as npt
 
-from pyvale.vfm.spatial_parameterisations.degree_of_freedom import (
-    DegreeOfFreedom,
-)
+from pyvale.vfm.constparam import ConstitutiveParameter
+from pyvale.vfm.dof import DegreeOfFreedom
+from pyvale.vfm.spatialparam import ISpatialParameterisation
 
 
 # TODO: maybe scrap this and use duck typing with an ORd type
 class IBasisFunctionKernel(ABC):
-    @property
     @abstractmethod
-    def num_degrees_of_freedom(self) -> int:
+    def get_num_degrees_of_freedom(self) -> int:
         pass
 
     @abstractmethod
@@ -40,8 +39,7 @@ class UnivariateBasisFunctionKernel(IBasisFunctionKernel):
     height: float | DegreeOfFreedom
     variance: float | DegreeOfFreedom
 
-    @property
-    def num_degrees_of_freedom(self) -> int:
+    def get_num_degrees_of_freedom(self) -> int:
         num_dofs = 0
 
         if isinstance(self.x, DegreeOfFreedom):
@@ -123,8 +121,7 @@ class BivariateBasisFunctionKernel(IBasisFunctionKernel):
     variance_y: float | DegreeOfFreedom
     angle: float | DegreeOfFreedom
 
-    @property
-    def num_degrees_of_freedom(self) -> int:
+    def get_num_degrees_of_freedom(self) -> int:
         num_dofs = 0
 
         if isinstance(self.x, DegreeOfFreedom):
@@ -221,3 +218,112 @@ class BivariateBasisFunctionKernel(IBasisFunctionKernel):
                 self.angle.value = dof.value
             else:
                 self.angle.value = dof
+
+
+# Initialising a new basis function:
+#   need to take:
+#   prev parameter map
+# TODO: need a global floor term dof
+@dataclass(slots=True)
+class BasisFunctionSpatialParameterisation(ISpatialParameterisation):
+    floor: float | DegreeOfFreedom | None = None
+    kernels: list[IBasisFunctionKernel] = field(default_factory=list)
+
+    def get_num_degrees_of_freedom(self) -> int:
+        num_dofs = 0
+
+        if isinstance(self.floor, DegreeOfFreedom):
+            num_dofs += 1
+
+        for kernel in self.kernels:
+            num_dofs += kernel.get_num_degrees_of_freedom()
+
+        return num_dofs
+
+    # TODO: create our initial basis functions with fitting
+    def update_from_constitutive_parameter(
+        self,
+        constitutive_parameter: ConstitutiveParameter
+    ) -> None:
+        initial_kernel_count = 1
+
+        
+
+    def to_map(
+        self,
+        size: npt.NDArray[np.uint32]
+    ) -> npt.NDArray[np.float64]:
+        ...
+        # map = np.zeros((size[0], size[1]))
+
+        # for kernel in self.kernels:
+            # dx = x - kernel.x_dof.value
+            # dy = y - kernel.y_dof.value
+
+            # variance_x = max(kernel.variance_x_dof.value, 1.0e-12)
+            # variance_y = variance_x
+            # angle = 0.0
+
+            # if kernel.variance_y_dof is not None:
+            #     variance_y = max(kernel.variance_y_dof.value, 1.0e-12)
+            # if kernel.angle_dof is not None:
+            #     angle = kernel.angle_dof.value
+            #     if angle_units.lower().startswith("deg"):
+            #         angle = np.deg2rad(angle)
+
+            # cos_theta = np.cos(angle)
+            # sin_theta = np.sin(angle)
+
+            # # Rotate coordinates into the kernel's principal directions.
+            # local_x = cos_theta * dx + sin_theta * dy
+            # local_y = -sin_theta * dx + cos_theta * dy
+
+            # exponent = -0.5 * (
+            #     (local_x * local_x) / variance_x
+            #     + (local_y * local_y) / variance_y
+            # )
+
+            # coefficient = 1.0
+            # if normalised:
+            #     determinant = variance_x * variance_y
+            #     coefficient = 1.0 / np.sqrt(determinant * (2.0 * np.pi) ** 2)
+
+            # return coefficient * kernel.height_dof.value * np.exp(exponent)
+
+    def collect_degrees_of_freedom(
+        self,
+    ) -> list[DegreeOfFreedom]:
+        dofs = []
+
+        if isinstance(self.floor, DegreeOfFreedom):
+            dofs.append(copy(self.floor))
+
+        for kernel in self.kernels:
+            dofs.append(kernel.collect_degrees_of_freedom())
+
+        return dofs
+
+    def update_from_degrees_of_freedom(
+        self,
+        degrees_of_freedom: list[DegreeOfFreedom] | npt.NDArray[np.float64]
+    ) -> None:
+        index = 0
+
+        if isinstance(self.floor, DegreeOfFreedom):
+            dof = degrees_of_freedom[index]
+
+            if isinstance(dof, DegreeOfFreedom):
+                self.floor.value = dof.value
+            else:
+                self.floor.value = dof
+
+            index += 1
+
+        for kernel in self.kernels:
+            num_dofs = kernel.get_num_degrees_of_freedom()
+
+            dofs = degrees_of_freedom[index:index + num_dofs]
+
+            kernel.update_from_degrees_of_freedom(dofs)
+
+            index += num_dofs
