+++ src/pyvale/vfm/spatialparamknown.py
@@ -3,23 +3,16 @@ from dataclasses import dataclass
 import numpy as np
 import numpy.typing as npt
 
-from pyvale.vfm.constitutive_laws.constitutive_parameter import (
-    ConstitutiveParameter,
-)
-from pyvale.vfm.spatial_parameterisations.degree_of_freedom import (
-    DegreeOfFreedom,
-)
-from pyvale.vfm.spatial_parameterisations.spatial_parameterisation import (
-    ISpatialParameterisation,
-)
+from pyvale.vfm.constparam import ConstitutiveParameter
+from pyvale.vfm.dof import DegreeOfFreedom
+from pyvale.vfm.spatialparam import ISpatialParameterisation
 
 
-@dataclass(slots=True, init=False)
+@dataclass(slots=True)
 class KnownSpatialParameterisation(ISpatialParameterisation):
     value: npt.NDArray[np.float64] | None = None
 
-    @property
-    def num_degrees_of_freedom(self) -> int:
+    def get_num_degrees_of_freedom(self) -> int:
         return 0
 
     def update_from_constitutive_parameter(
